fix(test): remove/fix vitest ESM warnings
```
(!) Your Vite config uses features that are unsupported by `configLoader: 'native'`, which is planned to become the default in a future major version of Vite:
  - import "../../vitest.config" without a file extension (vite.config.ts:8:32). Add the file extension
  - ESM syntax in a file loaded as CommonJS (../../vitest.config.ts:3:1). Use a `.mjs` extension or set `"type": "module"` in the closest package.json
Set `VITE_CONFIG_NATIVE_IGNORE_WARNING=true` to suppress this warning.
```
