+++ examples/medplum-client-external-idp-demo/vite.config.ts
@@ -5,8 +5,11 @@ import { copyFileSync, existsSync } from 'fs';
 import path from 'path';
 import { defineConfig } from 'vite';
 
-if (!existsSync(path.join(__dirname, '.env'))) {
-  copyFileSync(path.join(__dirname, '.env.defaults'), path.join(__dirname, '.env'));
+if (!existsSync(path.join(import.meta.dirname, '.env'))) {
+  copyFileSync(
+    path.join(import.meta.dirname, '.env.defaults'),
+    path.join(import.meta.dirname, '.env'),
+  );
 }
 
 export default defineConfig({
