+++ packages/app/vite.config.ts
@@ -6,11 +6,11 @@ import { execSync } from 'child_process';
 import { copyFileSync, existsSync } from 'fs';
 import path from 'path';
 import { defineConfig } from 'vite';
-import { medplumAliases } from '../../vitest.config';
+import { medplumAliases } from '../../aliases.mjs';
 import packageJson from './package.json' with { type: 'json' };
 
-if (!existsSync(path.join(__dirname, '.env'))) {
-  copyFileSync(path.join(__dirname, '.env.defaults'), path.join(__dirname, '.env'));
+if (!existsSync(path.join(import.meta.dirname, '.env'))) {
+  copyFileSync(path.join(import.meta.dirname, '.env.defaults'), path.join(import.meta.dirname, '.env'));
 }
 
 let gitHash;
@@ -40,9 +40,9 @@ export default defineConfig({
   resolve: {
     alias: {
       ...medplumAliases,
-      '@medplum/core': path.resolve(__dirname, '../core/src'),
-      '@medplum/react': path.resolve(__dirname, '../react/src'),
-      '@medplum/react-hooks': path.resolve(__dirname, '../react-hooks/src'),
+      '@medplum/core': path.resolve(import.meta.dirname, '../core/src'),
+      '@medplum/react': path.resolve(import.meta.dirname, '../react/src'),
+      '@medplum/react-hooks': path.resolve(import.meta.dirname, '../react-hooks/src'),
     },
   },
 });
